requirements.txt
numpy>=1.21.0
opencv-python>=4.5.0
scikit-image>=0.18.0
matplotlib>=3.4.0
Pillow>=8.3.0
tqdm>=4.62.0
joblib>=1.0.0
